Fixed timers in upload
There was a bug that the uploaders may never catch up because they
were pausing even if multiple files were waiting in the
queue to be uploaded. This problem could be compounded if the
plugin attempted a recovery and started with many items in the
queue.

The solution is to only delay if we have no pending uploads,
otherwise we keep uploading.