obsolete -> deprecated
Deprecate rather than obsolete config options so people migrating with ADC don't get affected with a broken pipeline.

The only obsolete feature is `key_path` because we now only accept the `json_key_file` or Application Default Credentials (ADC). This is marked obsolete because if we just started ignoring it then applications could start silently acting with the privileges of the host machine rather than what they were given with the key. The host machine might have an entirely different set of permissions or even be in a different project than the what the key is for.